[CLN]: split up log impls into separate files
